phildle-backend/requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataparser==0.0.2
dateparser==1.2.2
dotenv==0.9.9
Flask==3.1.1
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
html2text==2025.4.15
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
lxml==4.9.4
MarkupSafe==3.0.2
mwparserfromhell==0.7.2
psycopg2-binary==2.9.10
pycurl==7.45.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.4
six==1.17.0
slacker==0.14.0
soupsieve==2.7
SQLAlchemy==2.0.42
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.4.0
urllib3==2.5.0
Werkzeug==3.1.3
wikipedia==1.4.0
wikiquote==0.1.17
wikiquotes==1.4
wptools==0.4.17
